Removing test made on assumptions
Removing a test based on the assumption that iot products may not work as iotmq.  This broke because D600 was added to the list of iotmq devices, but we can't assume the other devices from the productiotmap may not be added to the iotmq list in the future and this will just keep breaking.  It really doesn't provide much value from a test perspective.